A LOT of changes - added fighter endpoints & other layers now registering with my Eureka server added service layer and dao layer template interfaces

+++ src/main/java/com/chris/accountservice/services/impl/AccountServiceImpl.java
@@ -1,12 +1,9 @@
 package com.chris.accountservice.services.impl;
 
-import com.chris.accountservice.controllers.AccountController;
 import com.chris.accountservice.dao.AccountDao;
 import com.chris.accountservice.dao.AccountLogDao;
-import com.chris.accountservice.exceptions.AccountNotFoundException;
 import com.chris.accountservice.models.Account;
 import com.chris.accountservice.models.AccountInfoDto;
-import com.chris.accountservice.models.AccountLog;
 import com.chris.accountservice.services.AccountService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -15,7 +12,6 @@ import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
 import javax.validation.Valid;
-import java.time.LocalDateTime;
 import java.util.*;
 
 
@@ -26,19 +22,12 @@ public class AccountServiceImpl implements AccountService {
     private AccountDao accountDao;
     @Autowired
     private AccountLogDao accountLogDao;
-    private static final Logger logger = LoggerFactory.getLogger(AccountController.class);
+    private static final Logger logger = LoggerFactory.getLogger(AccountServiceImpl.class);
 
-    private void longProcess() {
-        try {
-            Thread.sleep(10000);
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-            Thread.currentThread().interrupt();
-        }
-    }
+
     @Override
     public Set<Account> findAll() {
-        List<Account> accounts = accountDao.getAccounts();
+        List<Account> accounts = accountDao.get();
         if (!accounts.isEmpty()) {
             return new HashSet<>(accounts);
         } else {
@@ -46,35 +35,41 @@ public class AccountServiceImpl implements AccountService {
         }
     }
 
+
     @Override
     public Optional<Account> findById(Long id) {
-        return id == null ? Optional.empty() : accountDao.findById(id);
+        return id == null ? Optional.empty() : accountDao.getById(id);
     }
 
     @Override
     @Transactional
-    public Account save(Account account) {
-        account = accountDao.save(account);
+    public Account save(@Valid Account account) {
+        if (null != account) {
+            account = accountDao.save(account);
+        }
 
-        AccountLog accountLog = new AccountLog();
-        accountLog.setOperationType("Create new account");
-        accountLog.setTimestamp(LocalDateTime.now());
-        accountLog.setAccount(account);
-        accountLogDao.save(accountLog);
+//
+//        AccountLog accountLog = new AccountLog();
+//        accountLog.setOperationType("Create new account");
+//        accountLog.setTimestamp(LocalDateTime.now());
+//        accountLog.setAccount(account);
+//        accountLogDao.save(accountLog);
         return account;
     }
 
     @Override
-    public Account update(@Valid Account account) throws AccountNotFoundException {
+    public Account update(@Valid Account account) {
         account = accountDao.update(account);
         return account;
 
     }
 
     @Override
-    public void delete(Long id) throws AccountNotFoundException {
-        accountDao.delete(id);
-
+    public int delete(Long id) {
+        if (null == id) {
+            return 0;
+        }
+        return accountDao.delete(id);
     }
 
     @Override
@@ -84,4 +79,6 @@ public class AccountServiceImpl implements AccountService {
 
         return Set.copyOf(accountInfoDtos);
     }
+
+
 }
